requirements-dev.txt
-r requirements.txt
ipdb==0.10.1
flake8==3.0.4
model-mommy==1.3.2
pytest-cov==2.2.1
pytest-mock==1.2
pytest-django==3.1.2
